requirements.txt
alpha-vantage==2.1.1
aniso8601==8.0.0
attrs==19.3.0
certifi==2019.9.11
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.24.0
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
MarkupSafe==1.1.1
mongo==0.2.0
mongoengine==0.18.2
more-itertools==7.2.0
PyJWT==1.7.1
pymongo==3.9.0
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
Werkzeug==0.16.0
zipp==0.6.0
